Fix diazo error when fetching more events.
The problem is that diazo (even when disabled) tries to parse the response body.
If the body is empty (no more events), diazo throws an error while trying to access
the document root node.
By stripping the HTML, the content is an empty string and diazo tries no longer to parse it.