No longer use UID but activity record ID for batching.
Because we have change to the souper catalog, an object now may occure multiple times
in the activity feed.
Because of that, the UID of the object is not a good value for using in the batching.
Therefore we change to be based on the activity record id (intid) added by souper.