Fix bug with collective.geo when indexing objects too early.
The problem is that collective.geo may fire a IObjectGeoreferencedEvent
while creating the object, before the object is added to its parent.

Since the IObjectGeoreferencedEvent is also an IObjectModifiedEvent, we
are trying to add an activity record but fail because the acquisition is
missing.

The fix is to ignore all objects which have no acquisition.